+++ tests/test_utils.py
@@ -8,11 +8,13 @@ Specific test class:
    python -m unittest tests.utils.TestTicker
 
 """
-# import pandas as pd
+from datetime import datetime
+from unittest import TestSuite
+
+import pandas as pd
 # import numpy as np
 
-from .context import yfinance as yf
-from .context import session_gbl
+from tests.context import yfinance as yf
 
 import unittest
 # import requests_cache
@@ -35,16 +37,16 @@ class TestCache(unittest.TestCase):
         tkr = 'AMZN'
         tz1 = "America/New_York"
         tz2 = "London/Europe"
-        cache = yf.utils.get_tz_cache()
+        cache = yf.cache.get_tz_cache()
         cache.store(tkr, tz1)
         cache.store(tkr, tz2)
 
     def test_setTzCacheLocation(self):
-        self.assertEqual(yf.utils._DBManager.get_location(), self.tempCacheDir.name)
+        self.assertEqual(yf.cache._TzDBManager.get_location(), self.tempCacheDir.name)
 
         tkr = 'AMZN'
         tz1 = "America/New_York"
-        cache = yf.utils.get_tz_cache()
+        cache = yf.cache.get_tz_cache()
         cache.store(tkr, tz1)
 
         self.assertTrue(os.path.exists(os.path.join(self.tempCacheDir.name, "tkr-tz.db")))
@@ -61,10 +63,10 @@ class TestCacheNoPermission(unittest.TestCase):
         tz1 = "America/New_York"
 
         # During attempt to store, will discover cannot write
-        yf.utils.get_tz_cache().store(tkr, tz1)
+        yf.cache.get_tz_cache().store(tkr, tz1)
 
         # Handling the store failure replaces cache with a dummy
-        cache = yf.utils.get_tz_cache()
+        cache = yf.cache.get_tz_cache()
         self.assertTrue(cache.dummy)
         cache.store(tkr, tz1)
 
@@ -72,19 +74,43 @@ class TestCacheNoPermission(unittest.TestCase):
         # Test that if cache path in read-only filesystem, no exception.
         tkr = 'AMZN'
         # During attempt to lookup, will discover cannot write
-        yf.utils.get_tz_cache().lookup(tkr)
+        yf.cache.get_tz_cache().lookup(tkr)
 
         # Handling the lookup failure replaces cache with a dummy
-        cache = yf.utils.get_tz_cache()
+        cache = yf.cache.get_tz_cache()
         self.assertTrue(cache.dummy)
         cache.lookup(tkr)
 
 
+class TestPandas(unittest.TestCase):
+    date_strings = ["2024-08-07 09:05:00+02:00", "2024-08-07 09:05:00-04:00"]
+
+    @unittest.expectedFailure
+    def test_mixed_timezones_to_datetime_fails(self):
+        series = pd.Series(self.date_strings)
+        series = series.map(pd.Timestamp)
+        converted = pd.to_datetime(series)
+        self.assertIsNotNone(converted[0].tz)
+
+    def test_mixed_timezones_to_datetime(self):
+        series = pd.Series(self.date_strings)
+        series = series.map(pd.Timestamp)
+        converted = pd.to_datetime(series, utc=True)
+        self.assertIsNotNone(converted[0].tz)
+        i = 0
+        for dt in converted:
+            dt: datetime
+            ts: pd.Timestamp = series[i]
+            self.assertEqual(dt.isoformat(), ts.tz_convert(tz="UTC").isoformat())
+            i += 1
+
+
 def suite():
-    suite = unittest.TestSuite()
-    suite.addTest(TestCache('Test cache'))
-    suite.addTest(TestCacheNoPermission('Test cache no permission'))
-    return suite
+    ts: TestSuite = unittest.TestSuite()
+    ts.addTest(TestCache('Test cache'))
+    ts.addTest(TestCacheNoPermission('Test cache no permission'))
+    ts.addTest(TestPandas("Test pandas"))
+    return ts
 
 
 if __name__ == '__main__':